refactor(p1): extract WebMountTabsTools.kt + WebMountCaptureTools.kt — 4 WebMount tools
Two sibling factory files following the established pattern:
- WebMountTabsTools.kt: createTabListTool(deps) / createTabNewTool(deps) /
  createTabCloseTool(deps). Each takes only WebMountDeps because tabs
  operate purely on pool/activityStore.
- WebMountCaptureTools.kt: createScreenshotTool(deps). Same deps shape;
  WebViewScreenshot helper lives in primitives/ and isn't a ctor dep.

Coordinator replaces 4 `private val xxxTool = Tool(...)` fields with
`by lazy { createXxxTool(deps) }` delegators.

Verification: :app:compileDebugKotlin --offline → BUILD SUCCESSFUL.
