+++ app/src/main/java/me/rerere/rikkahub/data/agent/webmount/tools/WebMountPrimitiveTools.kt
@@ -944,158 +944,13 @@ class WebMountPrimitiveTools(
 
     // ----------------------------------------------------------- wm_tab_*
 
-    private val tabListTool = Tool(
-        name = "wm_tab_list",
-        description = """
-            List every live WebMount session in the pool with {session_id, url, title, status}.
-            Use this before wm_tab_close to see which sessions are open and what they hold.
-        """.trimIndent().replace("\n", " "),
-        parameters = {
-            InputSchema.Obj(properties = buildJsonObject {})
-        },
-        execute = { _ ->
-            deps.track("wm_tab_list", "WebMount 列出会话", buildJsonObject {}) {
-                val sessions = pool.listSessions().map { handle ->
-                    val ls = handle.loadState.value
-                    buildJsonObject {
-                        put("session_id", handle.sessionId)
-                        put("url", ls.currentUrl ?: ls.requestedUrl)
-                        put("title", ls.title)
-                        put("status", ls.status.wireName)
-                        put("destroyed", handle.destroyed)
-                    }
-                }
-                val payload = buildJsonObject {
-                    put("count", sessions.size)
-                    put("sessions", buildJsonArray { sessions.forEach { add(it) } })
-                }
-                listOf(UIMessagePart.Text(payload.toString()))
-            }
-        },
-    )
-
-    private val tabNewTool = Tool(
-        name = "wm_tab_new",
-        description = """
-            Allocate a brand-new WebMount session without navigating. Returns the freshly-issued
-            session_id. The session is empty (about:blank) — the agent typically follows with
-            wm_open to load the first URL. Useful when the agent wants to work on two pages in
-            parallel; tools on different sessions run concurrently (per-session parallel groups).
-        """.trimIndent().replace("\n", " "),
-        parameters = {
-            InputSchema.Obj(properties = buildJsonObject {})
-        },
-        execute = { _ ->
-            deps.track("wm_tab_new", "WebMount 新建会话", buildJsonObject {}) {
-                val handle = pool.acquireNew()
-                val payload = buildJsonObject {
-                    put("session_id", handle.sessionId)
-                    put("status", handle.loadState.value.status.wireName)
-                }
-                listOf(UIMessagePart.Text(payload.toString()))
-            }
-        },
-    )
-
-    private val tabCloseTool = Tool(
-        name = "wm_tab_close",
-        description = """
-            Destroy a WebMount session and release its WebView. After this, the session_id is no longer
-            valid — subsequent calls referencing it fail. Use to free memory when the agent is done
-            with a long-running session; the pool also LRU-evicts automatically at its capacity.
-        """.trimIndent().replace("\n", " "),
-        parameters = {
-            InputSchema.Obj(
-                properties = buildJsonObject {
-                    put("session_id", stringProp("Session id to close."))
-                    put("reason", stringProp("Optional reason recorded in logs."))
-                },
-                required = listOf("session_id"),
-            )
-        },
-        // Closing its own session is not a security action — auto-approvable to
-        // avoid training users to click "Approve" reflexively on a confirmation
-        // that doesn't carry real risk. Reverses the M1.4 review over-correction.
-        execute = { input ->
-            deps.track("wm_tab_close", "WebMount 关闭会话", input) {
-                val sessionId = input.requiredString("session_id")
-                val reason = input.string("reason") ?: "agent requested"
-                pool.release(sessionId, reason)
-                val payload = buildJsonObject {
-                    put("session_id", sessionId)
-                    put("closed", true)
-                    put("reason", reason)
-                }
-                listOf(UIMessagePart.Text(payload.toString()))
-            }
-        },
-    )
+    private val tabListTool by lazy { createTabListTool(deps) }
+    private val tabNewTool by lazy { createTabNewTool(deps) }
+    private val tabCloseTool by lazy { createTabCloseTool(deps) }
 
     // ---------------------------------------------------------- wm_screenshot
 
-    private val screenshotTool = Tool(
-        name = "wm_screenshot",
-        description = """
-            Capture the current page of a WebMount session and return its base64-encoded image in
-            the tool result text (a `data:` URI). `full_page=false` (default) captures the visible
-            viewport; `full_page=true` scrolls + stitches up to ~16384px tall. Uses native Android
-            Bitmap rendering with a software-rendering fallback for hardware-accelerated WebViews.
-            Heads up: full-page PNGs are large; for full-page captures prefer format="jpeg" with
-            quality=75-85 to keep the agent's context manageable.
-        """.trimIndent().replace("\n", " "),
-        parameters = {
-            InputSchema.Obj(
-                properties = buildJsonObject {
-                    put("session_id", stringProp("Session id."))
-                    put("full_page", booleanProp("Stitch scrolled slices (default false)."))
-                    put("format", stringProp("'png' (default) | 'jpeg'."))
-                    put("quality", integerProp("JPEG quality 1-100, default 85. PNG ignores this."))
-                },
-                required = listOf("session_id"),
-            )
-        },
-        execute = { input ->
-            deps.track("wm_screenshot", "WebMount 截图", input) {
-                val sessionId = input.requiredString("session_id")
-                val handle = pool.peek(sessionId) ?: error("session not found: $sessionId")
-                val fullPage = input.boolean("full_page") ?: false
-                val format = when (input.string("format")?.lowercase()) {
-                    "jpeg", "jpg" -> WebViewScreenshot.Format.JPEG
-                    else -> WebViewScreenshot.Format.PNG
-                }
-                val quality = (input.long("quality") ?: 85L).coerceIn(1L, 100L).toInt()
-                val result = WebViewScreenshot.capture(handle, fullPage, format, quality)
-                when (result) {
-                    is WebViewScreenshot.Result.Success -> {
-                        // Inline the base64 in the Text payload — every provider's
-                        // tool-result serializer drops UIMessagePart.Image, so the
-                        // model never sees standalone Image parts on a tool return.
-                        // The wm_screenshot Tool's `outputBudgetChars` is bumped
-                        // (see ToolRegistry.outputBudgetChars) so the base64 fits.
-                        val mime = if (result.format == "jpeg") "image/jpeg" else "image/png"
-                        val payload = buildJsonObject {
-                            put("session_id", sessionId)
-                            put("ok", true)
-                            put("width", result.width)
-                            put("height", result.height)
-                            put("format", result.format)
-                            put("size_bytes", result.sizeBytes)
-                            put("full_page", fullPage)
-                            put("image_data_url", "data:$mime;base64,${result.base64}")
-                        }
-                        listOf(UIMessagePart.Text(payload.toString()))
-                    }
-                    is WebViewScreenshot.Result.Failed -> {
-                        listOf(UIMessagePart.Text(buildJsonObject {
-                            put("session_id", sessionId)
-                            put("ok", false)
-                            put("error", result.message)
-                        }.toString()))
-                    }
-                }
-            }
-        },
-    )
+    private val screenshotTool by lazy { createScreenshotTool(deps) }
 
     // -------------------------------------------------------- wm_stations
